refactor(loss_hub): move get_loss_function dispatcher to package __init__
Mirrors `rollout/rm_hub/__init__.py:async_rm` — the dispatcher belongs
at the package level, not at the bottom of `losses.py` (where it had
to reach across base_types / load_function / each loss impl).

`from miles.backends.training_utils.loss_hub import get_loss_function`
is now the public entry point. `loss_hub/losses.py` is reduced to
just the three loss implementations.
